+++ src/main/java/it/eng/idsa/businesslogic/usagecontrol/service/UsageControlService.java
@@ -15,7 +15,7 @@ public interface UsageControlService {
 	 * @return
 	 * @throws Exception
 	 */
-	public String enforceUsageControl(URI contractAgreementUri, String payload) throws Exception;
+	String enforceUsageControl(URI contractAgreementUri, String payload) throws Exception;
 
 	/**
 	 * Used on the Usage Control provider side for creating Usage Control Object
@@ -23,7 +23,7 @@ public interface UsageControlService {
 	 * @param payloadContent
 	 * @return
 	 */
-	public String createUsageControlObject(ArtifactRequestMessage artifactRequestMessage,
+	String createUsageControlObject(ArtifactRequestMessage artifactRequestMessage,
 			ArtifactResponseMessage artifactResponseMessage, String payloadContent);
 
 	/**
@@ -31,12 +31,19 @@ public interface UsageControlService {
 	 * @param payloadContent
 	 * @return
 	 */
-	public String uploadPolicy(String payloadContent);
+	String uploadPolicy(String payloadContent);
 
 	/**
 	 * Used to rollback policy upload if clearing house logging fails
 	 * @param contractAgreement
 	 * @throws IOException 
 	 */
-	public void rollbackPolicyUpload(String contractAgreement);
+	void rollbackPolicyUpload(String contractAgreement);
+	
+	/**
+	 * Check the availability of the usage control
+	 * @param usageContolHealthEndpoint
+	 * @return
+	 */
+	boolean isUsageControlAvailable(String usageContolHealthEndpoint);
 }