+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.1.23-SNAPSHOT] - 2022-11-18
+## [0.1.24-SNAPSHOT] - 2022-11-30
 
 ### Added
  - New properties for Clearing House
@@ -13,6 +13,33 @@ All notable changes to this project will be documented in this file.
 ### Removed
 
  - ENG Clearing House dependency removed
+ 
+## [0.1.23-SNAPSHOT] - 2022-11-25
+
+### Added
+
+ - Resource manager for calculating CPU and memory usage
+ - Logic for checking internal and external "health" of the connector.
+ Internal health: dataApp and Usage Control app reachability, audit file volume usage
+ External health: DAPS and Clearing House reachability.
+ Note: UC dataApp version required is +v1.4
+ New properties needed for this logic:
+ 
+```
+application.healthcheck.enabled=false
+application.healthcheck.cron.expression=0 */5 * ? * *
+## Time in miliseconds
+application.healthcheck.resourcemanager.cron.fixedDelay=10000
+## Threshold in percentages - from max volume size
+application.healthcheck.threshold.audit=5
+
+application.healthcheck.threshold.cpu=0.5
+application.healthcheck.threshold.memory=10.5
+
+application.healthcheck.dataapp=
+application.healthcheck.usagecontrol
+```
+ 
 
 ## [0.1.22-SNAPSHOT] - 2022-10-31